Refactor TTL extension calls in Escrow contract and remove unused ReclaimedEvent import in tests

+++ contracts/escrow/src/lib.rs
@@ -160,7 +160,7 @@ impl EscrowContract {
         env.storage().persistent().set(&key, &escrow);
         env.storage()
             .persistent()
-            .extend_ttl(&key, TTL_THRESHOLD as u32, ESCROW_TTL as u32);
+            .extend_ttl(&key, TTL_THRESHOLD, ESCROW_TTL);
     }
 
     /// Release funds from escrow to the artisan
@@ -202,7 +202,7 @@ impl EscrowContract {
         env.storage().persistent().set(&key, &escrow);
         env.storage()
             .persistent()
-            .extend_ttl(&key, TTL_THRESHOLD as u32, ESCROW_TTL as u32);
+            .extend_ttl(&key, TTL_THRESHOLD, ESCROW_TTL);
     }
 
     /// Allow the client to reclaim funds after the deadline has passed when an escrow is still funded.
@@ -244,7 +244,7 @@ impl EscrowContract {
         env.storage().persistent().set(&key, &escrow);
         env.storage()
             .persistent()
-            .extend_ttl(&key, TTL_THRESHOLD as u32, ESCROW_TTL as u32);
+            .extend_ttl(&key, TTL_THRESHOLD, ESCROW_TTL);
 
         // Emit event
         env.events().publish(
